requirements.txt
openai==1.35.7
python-dotenv==1.0.1